+++ src/integration/benchmark_test.go
@@ -637,6 +637,29 @@ func (self *IntegrationSuite) TestIssue105(c *C) {
 	c.Assert(data[0].Points[0][3], Equals, 1.0)
 }
 
+func (self *IntegrationSuite) TestWhereConditionWithExpression(c *C) {
+	err := self.server.WriteData(`
+[
+  {
+    "name": "test_issue_105",
+    "columns": ["time", "a", "b"],
+    "points":[
+		  [1386262529794, 2, 1],
+		  [1386262529794, 2, 0]
+	  ]
+  }
+]`, "time_precision=m")
+	c.Assert(err, IsNil)
+	bs, err := self.server.RunQuery("select a, b from test_issue_105 where a + b >= 3")
+	c.Assert(err, IsNil)
+	data := []*h.SerializedSeries{}
+	err = json.Unmarshal(bs, &data)
+	c.Assert(data, HasLen, 1)
+	c.Assert(data[0].Columns, HasLen, 4)
+	c.Assert(data[0].Points, HasLen, 1)
+	c.Assert(data[0].Points[0][3], Equals, 1.0)
+}
+
 // test for issue #41
 func (self *IntegrationSuite) TestDbDelete(c *C) {
 	err := self.server.WriteData(`